fix(medium): M27, M82, M84, M87 — ucsrc_reset, ExtInterrupt tick, LCD float, CpuControl double
M27: UART reset now uses desc_.ucsrc_reset (default 0x06) instead of
     hardcoded 0x06; default in UartDescriptor changed to 0x06
M82: ExtInterrupt tick() skips refresh_bound_input when elapsed_cycles == 0
M84: LCD frame timing uses explicit double casts for precision
M87: CpuControl effective_frequency() uses double instead of float

+++ src/core/cpu_control.cpp
@@ -86,8 +86,8 @@ u64 CpuControl::effective_frequency() const noexcept
         // 0x00 -> 50% frequency
         // 0x80 -> 100% frequency
         // 0xFF -> ~150% frequency
-        float scale = 0.5f + (static_cast<float>(osccal_) / 128.0f) * 0.5f;
-        freq = static_cast<u64>(static_cast<float>(freq) * scale);
+        double scale = 0.5 + (static_cast<double>(osccal_) / 128.0) * 0.5;
+        freq = static_cast<u64>(static_cast<double>(freq) * scale);
     }
     
     // 2. Apply CLKPR prescaler
